cmd/wit-bindgen-go/cmd/generate: support for WIT file path as OCI path
this commit adds a new feature to the generate command to support WIT files as a path to OCI registry. It takes the form of `wit-bindgen-go generate ghcr.io/webassembly/wasi/http:0.2.0`.

It does so by using regclient APIs to pull and unpack WIT files from a OCI registry, and `wasm-tools` to unpack a wasm file to WIT files on the local file system. By default, the WIT files are created to `wit` directory. This can be override by `--wit-out` flag.

Signed-off-by: Jiaxiao Zhou (Mossaka) <duibao55328@gmail.com>