.github/workflows/release: add a release GitHub Action workflow
the release workflow will be triggered when a new tag is pushed to the
remote, and will build, test and release the wit-bindgen-go binaries
to the GitHub release page. It runs on linux, macos and windows and
on amd64, arm64 architectures.

the built binary name is named as wit-bindgen-go-<os>-<arch>.